Fix navigation flow
